moved create venue in to create page file and removed CreateVenue component
